services/clinic_config_service/pyproject.toml
[project]
name = "clinic-config-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt==4.2.1",
    "fastapi>=0.121.2",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "aerich>=0.9.2",
    "deptry>=0.24.0",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.5",
]

[tool.aerich]
tortoise_orm = "app.db.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
